old/flask_and_postgres/requirements.txt
kafka-python==2.0.2
pymongo==4.3.3
SQLAlchemy==1.4.22
psycopg2-binary==2.9.1
Flask==2.0.1

